+++ tests/CSharpDotNetProject/src/CSharpDotNetProject/Constructors.cs
@@ -1,0 +1,13 @@
+namespace CSharpDotNetProject;
+
+public class PrimaryConstructorClass
+{
+    public int Int { get; }
+    public string String { get; }
+
+    public PrimaryConstructorClass(int Int, string String)
+    {
+        this.Int = Int;
+        this.String = String;
+    }
+}